rename js script to lib for now and move out of components
